implement AuthUser middleware

+++ src/app/api/admin/delete-product/route.js
@@ -1,4 +1,5 @@
 import connectToDB from "@/database";
+import AuthUser from "@/middleware/AuthUser";
 import Product from "@/models/product";
 import { NextResponse } from "next/server";
 
@@ -11,35 +12,45 @@ export const dynamic = 'force-dynamic';
 export const DELETE = async (req) => {
     try {
         await connectToDB();
-        const { searchParams } = new URL(req.url);
-        const id = searchParams.get('id');
+        const isAuthUser = AuthUser(req);
 
-        if (!id) {
-            return NextResponse.json({
-                success: false,
-                message: 'Product ID is required!'
-            });
-        }
+        if (isAuthUser?.admin === 'admin') {
 
-        const deletedProduct = await Product.findByIdAndDelete(id);
+            const { searchParams } = new URL(req.url);
+            const id = searchParams.get('id');
 
-        if (deletedProduct) {
-            return NextResponse.json({
-                success: true,
-                message: 'Product deleted successfully!'
-            });
+            if (!id) {
+                return NextResponse.json({
+                    success: false,
+                    message: 'Product ID is required!'
+                });
+            }
+
+            const deletedProduct = await Product.findByIdAndDelete(id);
+
+            if (deletedProduct) {
+                return NextResponse.json({
+                    success: true,
+                    message: 'Product deleted successfully!'
+                });
+            } else {
+                return NextResponse.json({
+                    success: false,
+                    message: 'Failed to delete a product!'
+                });
+            }
         } else {
             return NextResponse.json({
                 success: false,
                 message: 'Failed to delete a product!'
-            })
+            });
         }
 
     } catch (error) {
         console.log(error);
         return NextResponse.json({
             success: false,
-            message: 'Something went wrong! Please try again!'
-        })
+            message: 'You are not authenticated!'
+        });
     }
 }
